SonarLint: cast one of the operands of this subtraction operation to a 'double'

+++ SortingAnimations/src/java/example/MainPanel.java
@@ -109,7 +109,7 @@ public final class MainPanel extends JPanel {
   public void genArray(int n) {
     array.clear();
     factorx = (MAXX - MINX) / (double) n;
-    factory = MAXY - MINY;
+    factory = (double) MAXY - MINY;
     distributionsChoices.getItemAt(distributionsChoices.getSelectedIndex()).generate(array, n);
   }
